perf: replace 1000s of event listeners by a single one 💪
Every node in the file tree had an event listener for its various button. We replaced all of this event listeners by a single event listener, that deduces which node is clicked at the click moment, instead of adding an event listener to every single button. This way we go from thousands of event listeners to one single event listener for the whole tree view 💪